Added two concrete types to Pel: vector and matrix, both based on the  corresponding types from boost::numeric::ublas.
Also fixed the python/p2 library build rules.

+++ trunk/elements/csvparser.C
@@ -32,8 +32,8 @@
 //  - push-in will be reenabled when there are less than MIN_Q_SIZE
 //    tuples in the output queue. 
 //
-const size_t CSVParser::MIN_Q_SIZE = 100;
-const size_t CSVParser::MAX_Q_SIZE = 100;
+const uint32_t CSVParser::MIN_Q_SIZE = 100;
+const uint32_t CSVParser::MAX_Q_SIZE = 100;
 
 // 
 // Constructor.
